Remove logic for loading a user's editor settings
Atom v1.25.0 changed how user configs were updated on disk, subsequently
removing some properties which we were relying on to find `config.cson`.
This is causing `loadEditorSettings` to raise an exception at startup.

Since this feature was headed for the incinerator anyway, we may as well
cull it early (it was a stupid idea to begin with, among many others yet
to be obliterated).

See Alhadis/Atom-Mocha#3.