Give screenshot specs ample time to finish running
Taking a screen-capture of the host's desktop can take up to ~12 seconds
to complete on Windows (when run by GitHub Actions, at least). This is a
steep increase from the ~2-3 seconds it takes for node(1) to execute the
same code in the same environment.

Also, make sure we're saving the capture to a directory that exists.